+++ test/mx_http_test.dart
@@ -1,1 +1,9 @@
-void main() {}
+import 'package:test/test.dart';
+
+import '../example/mx_http_example.dart' as ex;
+
+void main() {
+  test('test1', () async {
+    await ex.main();
+  });
+}